requirements.txt
pandas 
mysql-connector-python 
sqlalchemy